settings restructuring
