+++ ArtsyFolio Tests/View Controllers/ARLabSettingsViewControllerTests.m
@@ -1,6 +1,5 @@
 #import "ARLabSettingsSplitViewController.h"
 #import "ARStoryboardIdentifiers.h"
-#import "ARLabSettingsDetailViewManager.h"
 #import "ARSyncStatusViewModel.h"
 #import "SyncLog.h"
 #import <ISO8601DateFormatter/ISO8601DateFormatter.h>
@@ -15,8 +14,6 @@ __block NSManagedObjectContext *context;
 __block ARSyncStatusViewModel *syncViewModel;
 
 beforeEach(^{
-    ARLabSettingsDetailViewManager *manager = [[ARLabSettingsDetailViewManager alloc] init];
-    mockManager = [OCMockObject partialMockForObject:manager];
     context = [CoreDataManager stubbedManagedObjectContext];
     
     SyncLog *syncLog = [SyncLog objectInContext:context];
@@ -24,7 +21,6 @@ beforeEach(^{
     syncLog.dateStarted = [formatter dateFromString:@"2015-10-31T02:22:22"];
     
     syncViewModel = [[ARSyncStatusViewModel alloc] initWithSync:nil context:context];
-    [[[mockManager stub] andReturn:syncViewModel] viewModelForSection:ARLabSettingsSectionSync];
 });
 
 afterEach(^{
