--- Classes/Controllers/Settings/ARLabSettingsDetailViewManager.h
@@ -1,19 +1,0 @@
-typedef NS_ENUM(NSInteger, ARLabSettingsSection) {
-    ARLabSettingsSectionSync,
-    ARLabSettingsSectionPresentationMode,
-    ARLabSettingsSectionBackground,
-    ARLabSettingsSectionEmail,
-};
-
-@protocol SettingsSectionViewController
-
-@end
-
-
-@interface ARLabSettingsDetailViewManager : NSObject <UISplitViewControllerDelegate>
-
-@property (nonatomic, strong) NSManagedObjectContext *context;
-
-- (id)viewModelForSection:(ARLabSettingsSection)section;
-
-@end
