+++ Classes/Util/App/ARStoryboardIdentifiers.h
@@ -2,8 +2,12 @@
 
 #import <Foundation/Foundation.h>
 
+#pragma mark - navigationController.storyboardIdentifier
+extern NSString *const SettingsNavigationController;
+
 #pragma mark - segue.identifier
-extern NSString *const SyncSettings;
+extern NSString *const ShowPresentationModeSettingsViewController;
+extern NSString *const ShowSyncSettingsViewController;
 
 #pragma mark - storyboardNames
 extern NSString *const ARLabSettings;
@@ -12,5 +16,6 @@ extern NSString *const ARLabSettings;
 extern NSString *const previousSyncCell;
 
 #pragma mark - viewController.storyboardIdentifier
+extern NSString *const PresentationModeViewController;
 extern NSString *const SettingsMasterViewController;
-extern NSString *const SyncViewController;
+extern NSString *const SyncSettingsViewController;
