+++ Classes/Util/App/ARStoryboardIdentifiers.m
@@ -2,8 +2,12 @@
 
 #import "ARStoryboardIdentifiers.h"
 
+#pragma mark - navigationController.storyboardIdentifier
+NSString *const SettingsNavigationController = @"SettingsNavigationController";
+
 #pragma mark - segue.identifier
-NSString *const SyncSettings = @"Sync Settings";
+NSString *const ShowPresentationModeSettingsViewController = @"Show Presentation Mode Settings View Controller";
+NSString *const ShowSyncSettingsViewController = @"Show Sync Settings View Controller";
 
 #pragma mark - storyboardNames
 NSString *const ARLabSettings = @"ARLabSettings";
@@ -12,5 +16,6 @@ NSString *const ARLabSettings = @"ARLabSettings";
 NSString *const previousSyncCell = @"previousSyncCell";
 
 #pragma mark - viewController.storyboardIdentifier
+NSString *const PresentationModeViewController = @"Presentation Mode View Controller";
 NSString *const SettingsMasterViewController = @"Settings Master View Controller";
-NSString *const SyncViewController = @"Sync View Controller";
+NSString *const SyncSettingsViewController = @"Sync Settings View Controller";
